Consolidate statistics merging code
